Clarify return type behaviour
It's not obvious from this documentation that when creating mocks, if you specify a return type then the mock will return that type. This is most prominent when upgrading code to supply type-hints, as doing so may cause false-positives in test cases which are relying ``null`` to be returned if not defined otherwise.